// src/users/enums/user-permissions.enum.ts
export enum userPermissionsEnum {
  CREATE = "ایجاد کاربر",
  LIST = "دریافت لیست کاربران",
  FIND_ONE = "دریافت یک کاربر",
  UPDATE = "به روز رسانی اطلاعات کاربر",
  DELETE = "حذف کاربر",
  ACCESS_CHECK = "بررسی دسترسی کاربر",
  ADD_ROLE = "اضافه کردن نقش به کاربر",
  REMOVE_ROLE = "حذف نقش کاربر",
  HARD_DELETE = "حذف عمیق کاربر",
  SET_AMOUNT = "set-amount-user",
  GET_AMOUNT = "get-amount-user",
}